// src/shared/interfaces/student-enrollment.interface.ts
export interface IStudentEnrollmentData {
  Usuario: string | null;
  Nombre: string | null;
  Apellido: string | null;
  Email: string | null;
  Curso: string | null;
  tipo_curso: string | null;
  Fecha_Matricula: string | null;
  Anio: number | null;
  Mes: number | null;
  Dia: number | null;
}